No
We developed a novel approach named alpha-HClustering (alpha-HC) for community detection in social networks. Our work is based on an extended similarity named alpha-structural-similarity(alpha-2S). Then we extended the HC named alpha-HC. H_clustering_approach.py : is the implementation of Hc approach which uses the structural similarity ( same to salton smilarity sing otherwise defined neighborhood sets) benchmarking_algorithms.py : is the implementation of the other approaches in the paper h_clustering_new_Sim.py : is the implementation of alpha-HC